Do not override error msg's of CanBuildFrom #643
Addresses issue #643 which contained an example of a custom error
message being lost for a List type when an item in that list failed to
decode properly.

The problem is present for three basic types of collections

	- Set
	- List
	- Vector

since each of these decoders defined in Decoder.scala call
withErrorMessage and therefore lose any message bubbled up from within
that collection.