Transform Codec from type-alias to trait #133
This was done to allow the definition of an implicit def that creates
a Codec for some type 'A' if we have an Encoder + Decoder for it in scope.

This implicit def can't be put anywhere else because then nothing compiles
anymore due to implicit search divergence problems, and the only way to
put it into a "companion object" is if it's a top level type.

This can be reverted if it is deemed necessary that such an implicit def
is undersireable.