Tests: migrate from ScalaTest to MUnit. (#1622)
* Migrate Tests from ScalaTest to MUnit.

Following a common policy for other projects, we migrate tests from
the scalatest library to the munit testing library, because the later
is smaller, leaner and has fewer Scala dependencies, which helps
its release and upgrade cycles.

We introduce a `CirceMunitSuite` trait, that parallels
`CirceSuite`. For most test suites, we just need to switch traits
and then rewrite tests into munit `test` or `property`.

For KeyEncoderSuite, KeyDecoderSuite, and KeyCodecSuite, and
 most of the CodecSuites, we only need to switch the suite trait.

- Array Seq Suite...
- LiteralCodecSuite...
- JawnParserSuite..
- DerivesSuite...
- JavaTimeCodecSuite
- Scodec integration
- Extras
- Refined
- Shapeless
- PrinterWriterReuseSuite
- SerializableSuite
- Generic

* Undo changes to JsonObjectSuite.

These tests are breaking at some poing, because of a corner case
with generators. We are undoing changes here to let the rest of
the build succeed...

* Generics for Scala-2: migrate tests scalatest -> munit.

Use smaller parts