ReadCursor: introduce superclass of `ACursor` for read-only acitons
The `ACursor` class provides three types of methods: read operations,
to get the information under the cursor, navigation, to move the
cursor up and down or sideways in the Json AST it is decoding,
and modifications.

However, it seems to us that Decoding should not require the use of
those modification operations. This is important if we want to
support decoding from fixed representations, such as byte arrays.