Derive Decoder and Codec under -Yexplicit-nulls
Fixes #1786.

Since this requires a new compiler flag, I had to come up with some way
to test it. I looked into setting this flag for the tests in the tests
module, but this would have made a lot of code that is shared with
Scala 2 stop compiling, including where JDK APIs return values that are
never null but not annotated as such. I opted instead to add a
dependency on scala3-staging and use the "Runtime Multi-Stage
Programming" describe at
https://dotty.epfl.ch/docs/reference/metaprogramming/staging.html