Permit Parsing Of Leading Zeros
This commit changes `BiggerDecimal.parseBiggerDecimalUnsafe` to permit parsing of leading '0' characters, and by extension allows all of our numeric codecs to tolerate decoding `JsonString` values which represent numbers with leading zeros. This is more in line with the `fromString` methods on the various numeric classes.

Note, this change should be given some serious scrutiny. It appears that the original authors had at some time explicitly forbade this behavior, as noted by the existence of the `AFTER_ZERO` state as well as the explicit test case for invalid input including the `String` `"01"`.
p
I've searched around the issues and I couldn't find any reasoning for this decision, so I still think being lenient here is the right choice, but please double check me.