Fix The PointerInterpolatorSuite For All Scala Versions
This takes the fix from https://github.com/circe/circe/pull/1975 and ports it to all Scala versions. At the time I wrote the original fix, I didn't realize we had this class split for the Scala 2/3.

To fix this issue for all Scala versions, I added `genPointerReferenceString` which generates a valid JSON pointer reference value, properly handling the escape.

Note, there is probably still more to do here in the future. The tests are slightly different in their structure for Scala 2/3 and from glancing at the RFC I suspect there are issues with the current pointer implementation when a reference contains Unicode characters which are represented by surrogate pairs of bytes under UTF-8. However, fixing those would be a more intrusive change and thus reserved for a different PR.