Use ZoneOffset Rather Than ZoneId In Most Arbitraries
This works around special case rules for ZoneId values which cause issues with testing. We still test ZoneId itself, but we use ZoneOffset to construct the various time values.

This commit also change the `Arbitrary` `ZoneOffset` to be based of the ScalaCheck implementation which fully covers the domain. We can remove this once https://github.com/circe/circe/issues/1994 is complete.