Use The Jawn Parser For Parsing Sorted Output In Tests
Fixes #1911

The `JSON.parse` call on native JS automatically and always sorts the keys when parsing, but slightly differently than the ordering prescribed by `Order[String]`. This causes tests on the sorted printers to fail under certain circumstances.

The fix for this is to use Jawn on JS to get consistent parsing. If we switch to Jawn in general for both JS and the JVM then we can undo this.

Note, the issue has never been with the printing, but always with the parsing.