Handle FailedCursor Terminal State Better In Path Construct
When the cursor is a `FailedCursor` as a result of a failed JSON traversal, e.g. `downField("foo")`, then we need to inspect the `lastOp` to report the path we were _attempting_ to move to. There are a lot of odd states here due to how cursors are modeled in 0.14.x. We can do better in 0.15.x, but I don't think we can do much better on 0.14.x without breaking changes.