Fix Scalafix For Modules Which Only Support Scala2
Prior to this commit we were doing this to avoid using Scala 3 on the 3 modules which do not support it,

```scala
crossScalaVersions := (ThisBuild / crossScalaVersions).value.filterNot(_.startsWith("3."))
```

Warning, things start to get weird at this point.

This means that when the outer `root` module is at `3.1.0`, an inner module which is using this will be at `2.12` or `2.13`.

```
sbt:rootJS> show scalaVersion
[info] coreJS / scalaVersion
[info]  3.1.0
[info] scalajs / scalaVersion
[info]  3.1.0
[info] pointerLiteralJS / scalaVersion
[info]  3.1.0
[info] literalJS / scalaVersion
[info]  3.1.0
[info] numbersTestingJS / scalaVersion
[info]  3.1.0
[info] refinedJS / scalaVersion
[info]  3.1.0
[info] scalajsJavaTimeTest / scalaVersion
[info]  3.1.0
[info] hygieneJS / scalaVersion
[info]  2.13.8
[info] parserJS / scalaVersion
[info]  3.1.0
[info] jawnJS / scalaVersion
[info]  3.1.0
[info] scodecJS / scalaVersion
[info]  3.1.0
[info] extrasJS / scalaVersion
[info]  2.13.8
[info] testsJS / scalaVersion
[info]  3.1.0
[info] pointerJS / scalaVersion
[info]  3.1.0
[info] testingJS / scalaVersion
[info]  3.1.0
[info] shapesJS / scalaVersion
[info]  2.13.8
[info] genericJS / scalaVersion
[info]  3.1.0
[info] numbersJS / scalaVersion
[info]  3.1.0
[info] scalaVersion
[info]  3.1.0
```

Note that `hygieneJS`, `extrasJS`, and `shapesJS` are all at `2.13.8`, which the others are at `3.1.0`.

Even though these projects will not be built for Scala 3, they still _technically_ have their normal modules dependencies in the build for Scala 3. For example, `shapesJS` depends `coreJS` and thus even for Scala 3 `sbt` will demand that we can find the class files for `core`. In this case they will be Scala 3 class files, which is broken and won't work, but we won't care because they won't get built together or published.

```
sbt:rootJS> show shapesJS/dependencyClasspathFiles
[info] * /tmp/tmp.XLMclp0VGr/circe/modules/core/js/target/scala-3.1.0/classes
[info] * /tmp/tmp.XLMclp0VGr/circe/modules/numbers/js/target/scala-3.1.0/classes
```

However, once `dependsOn(scalafixInternalRules % ScalafixConfig)` was added something changed (I don't know what) and now `sbt` started looking for the classes for `core` and `numbers` at for Scala 2.13 at the current `SHAPSHOT` version, which doesn't exist when the outer version is set to Scala 3.

```
[error] (hygieneJS / update) sbt.librarymanagement.ResolveException: Error downloading io.circe:circe-core_sjs1_2.12:0.14.2-101-628559a-20220819T164808Z-SNAPSHOT
[error]   Not found
[error]   Not found
[error]   not found: /home/dstrawn/.ivy2/local/io.circe/circe-core_sjs1_2.12/0.14.2-101-628559a-20220819T164808Z-SNAPSHOT/ivys/ivy.xml
[error]   not found: https://repo1.maven.org/maven2/io/circe/circe-core_sjs1_2.12/0.14.2-101-628559a-20220819T164808Z-SNAPSHOT/circe-core_sjs1_2.12-0.14.2-101-628559a-20220819T164808Z-SNAPSHOT.pom
[error]   not found: https://repo.scala-sbt.org/scalasbt/sbt-plugin-releases/io.circe/circe-core_sjs1_2.12/0.14.2-101-628559a-20220819T164808Z-SNAPSHOT/ivys/ivy.xml
[error]   not found: https://repo.typesafe.com/typesafe/ivy-releases/io.circe/circe-core_sjs1_2.12/0.14.2-101-628559a-20220819T164808Z-SNAPSHOT/ivys/ivy.xml
[error] Error downloading io.circe:circe-literal_sjs1_2.12:0.14.2-101-628559a-20220819T164808Z-SNAPSHOT
```

Why does adding this `dependsOn` cause this? I've no idea.

The "fix" is to permit the `crossScalaVersions` to contain Scala 3 for these projects, but to `skip` all of their meaningful tasks, e.g. compile/test/publish. The result is the same, they are only built for Scala 2, not Scala 3, but the dependencies don't get messed up.