Normalize Module Name
This commit adds a new function to the `build.sbt` which normalizes the module name in the cases where the module name is derived from a path with a sub-directory, and thus will include a `'/'` character. These characters are replaced with `'-'`, which is both valid and probably want we intended in the first place.